Enabled testing of pattern in build log
It is now possible to provide a URL to a build log
in the text box used for matching against the pattern
in a build log indication. A drop down list is used
for selection of how to treat the text in the text
box - as a plain text or as a build URL. When a URL
is given, the matching is done against the log of
the build with the given URL. The first matching
line to be found, if any, is printed.

If the failure cause management page is entered
from a project page, the URL to the latest
failed build for the project is pre-written
in the text box, if the URL can be found.

Change-Id: Ia2cc4d5d91fd93bb8b7289211ce477c11b5000ee