Fix for serialization of the matrix aggregated build action
The issue is that MatrixRuns are there own item group and is not
serialized together with the main matrix build, the action contained
references to the runs so they got serialized into the main build
with the action.
When the action was deserialized those runs would become their own
separate object instances with no knowledge of the combination
they belonged to, wich would cause several strange behaviours.
We solve this by storing string id's of the run combinations and
find the "real" runs when called for.
But because that the runs are separate item groups they are not
yet available in the object graph when the action's readResolve
method is invoked after deserialization, so converting from run
instance to runId is a bit cumbersome to fix there.
So the conversion is performed in a separate thread pool.

Change-Id: I4b19b8c4b53551e68ff43fd91e4d7b9401316206