Replace desc. placeholders from captured expr*
Substitutions may be made within the description with
placeholders of the form ${I,G}, where I is the indication
number and G is the captured group within the indication
expression. e.g., ${1,1} would be replaced with the first
indication's first captured group and ${1,2} would be
replaced with the first indication's second captured group.