Pipeline support for Build Failure Analyzer
With this change, Pipeline projects are analyzed
by the BFA. The major change was changing AbstractBuild
to Run and AbstractProject to Job, which would also make
other project and build types that aren't AbstractBuild/Project
start working with the BFA.

Some issues remain and should be fixed in future changes:
Downstream builds linking. If a Pipeline project starts downstream builds
which fail, these won't be shown on the Pipeline build page.

"Built on". In the statistics logging, adding which node the Pipeline
run was run on doesn't work. A Pipeline project could use many nodes
and a new structure for this in the statistics must be created.

TokenMacro plugin integration. The TokenMacro evaluation uses AbstractBuild.
This needs to be changed to Run in the TokenMacro plugin in order for the
evaluation to work with Pipeline runs.

Change-Id: I2a0debed0a13c051d3614c5ceece77a31232ac3d