New Extension Point: IndicationFilter
IndicationFilter allows plugins to hide specific indications
from the CauseManagement UI.

Indications added before a Filter is applied will still continue
to run.

The main purpose of this ExtensionPoint is to allow plugin
contributed indications to hide core indications which supply
similiar (or just a subset) of the core indicators' functionality.

Change-Id: I0dc492cdedb1232850262e1d11db35009031d9a3