Fix failed tests
One testcase was a bit broken (it only worked due to a somewhat
faulty compiled pattern in previous versions of
MultilineBuildLogIndication). This test case was replaced with a
more stable one that utilizes the block timeout.

QuadrupleDupleLineReader() is not used for Multiline scanning
since it does not use Reader.readline().

Another testcase utilized a deprecated schedule method and did
not receive Causes due to this.

Change-Id: I6cf0dfc81869aad965c3d2731cfb8f9121da584f