Added a FailureCauseProvider for the MQ Notifier plugin
The MQ Notifier plugin sends messages to a Message Queue,
implementing the AMQP protocol, e.g. RabbitMQ, whenever a
build is finished.
The plugin allows for other plugins to add
information to the message. This commit adds support for
the Build Failure Analyzer to supply Failure Cause information
from the build.

The dependency to the MQ Notifier plugin is made optional.