Add a global switch to disable report when no causes identified
When a build fails and no build causes in the knowledge base are
matched against the console log, the build page reports with an
heading and a message that there was "No identified problem".

While most users would understand the message, we had reports
from the developers reporting that the message is very confusing,
in particular in the early stages after the plugin is deployed,
and the knowledge base does not include many frequent failures.

This patch introduces a global switch that lets the administrator
turn off the display of the build analyzer block on build pages
of jobs that failed and no indication was found.