New feature: generic causes
Allows to designate categories as generic. When scanning for
known causes, generic causes are only stored if there are no non-generic
causes present.

Use case:
- generic cause: A maven goal failed
- specific cause: The code did not compile

Old behaviour: Both causes would be shown, leading to descriptions like
"always look for more specific causes first"

New behaviour: if a compile error is found, only that one is shown
(maven error is ignored as generic). Otherwise, the information that
the failure was due to a maven error (along with the link!) is still
provided.