Include line number of the found indication
The line number is needed for allowing us to figure out which
indicator was found first.

Note: In this PR, the linenumber will only be provided for single
line indicators. While it may be feasible to add line counting to
the multiline indicators, it's a complemetely different beast. The
multiline implementation uses a sliding search window mechanism and
line counting will incur some refactoring. If ever needed, it should
be added in a future PR.